refactor(tx-builder): deprecate ID_TAG constants
